correçoes gerais de merge
